Code refactoring.

+++ src/back-end/Service/Catalog/Services/BaseService.cs
@@ -5,12 +5,12 @@ using Catalog.Models;
 
 namespace Catalog.Services
 {
-    public abstract class ServiceBase<T> : IService<T>
+    public abstract class BaseService<T> : IService<T>
         where T : IEntity
     {
         private readonly IRepository<T> _repository;
 
-        protected ServiceBase(IRepository<T> repository)
+        protected BaseService(IRepository<T> repository)
         {
             _repository = repository;
         }
@@ -21,8 +21,17 @@ namespace Catalog.Services
             return item;
         }
 
-        public async Task DeleteAsync(string id) =>
+        public async Task<bool> DeleteAsync(string id)
+        {
+            var item = await GetByIdAsync(id);
+            if (item == null)
+            {
+                return false;
+            }
+
             await _repository.DeleteAsync(id);
+            return true;
+        }
 
         public async Task<List<T>> GetAllAsync()
         {
@@ -39,7 +48,17 @@ namespace Catalog.Services
             return await _repository.ReadByIdAsync(id);
         }
 
-        public async Task UpdateAsync(string id, T itemIn) =>
+        public async Task<bool> UpdateAsync(string id, T itemIn) 
+        {
+            var item = await GetByIdAsync(id);
+            if (item == null || item.Id != itemIn.Id)
+            {
+                return false;
+            }
+
             await _repository.UpdateAsync(id, itemIn);
+            return true;
+        }
+            
     }
 }
